requirements.txt
Django>=1.8
django-bootstrap3-datetimepicker
python-mpd2
celery
sqlalchemy
